# docs/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Courier Delivery Management System
  description: "\n        ## Courier Delivery Management System API\n\n        A comprehensive\
    \ API for managing courier delivery operations with real-time tracking and updates.\n\
    \n        ### Features\n        - **User Management**: Authentication and role-based\
    \ access control\n        - **Package Management**: Create, track, and manage\
    \ delivery packages\n        - **Real-time Updates**: WebSocket connections for\
    \ live notifications\n        - **Role-based Access**: Admin, Courier, Sender,\
    \ and Recipient roles\n        - **Status Tracking**: Complete package lifecycle\
    \ management\n\n        ### Authentication\n        This API uses JWT (JSON Web\
    \ Tokens) for authentication. Include the token in the Authorization header:\n\
    \        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n\n\
    \        ### User Roles\n        - **Admin**: Full system access and management\n\
    \        - **Courier**: Manage assigned packages and update delivery status\n\
    \        - **Sender**: Create packages and track deliveries\n        - **Recipient**:\
    \ View packages addressed to them\n\n        ### Getting Started\n        1. Register\
    \ a new user account\n        2. Login to receive an access token\n        3.\
    \ Use the token to access protected endpoints\n        4. Connect to WebSocket\
...
    \        "
  contact:
...
  license:
...
servers:
...
paths:
  /api/auth/login:
    post:
      tags:
      - authentication
      summary: User Login
      description: Authenticate user with email and password to receive access token
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UserLogin'
              title: User Credentials
              description: User login credentials
            example:
              email: john.doe@example.com
              password: securepassword123
        required: true
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 1800
                user:
                  id: 1
                  email: john.doe@example.com
                  name: John Doe
                  role: sender
                  phone: '+1234567890'
                  created_at: '2024-01-01T10:00:00Z'
                  updated_at: '2024-01-01T10:00:00Z'
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
  /api/auth/register:
    post:
      tags:
      - authentication
      summary: User Registration
      description: Register a new user account
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UserCreate'
              title: User Data
              description: User registration data
            example:
              email: john.doe@example.com
              name: John Doe
              password: securepassword123
              role: sender
              phone: '+1234567890'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '201':
          description: User created successfully
          content:
            application/json:
              example:
                id: 1
                email: john.doe@example.com
                name: John Doe
                role: sender
                phone: '+1234567890'
                created_at: '2024-01-01T10:00:00Z'
                updated_at: '2024-01-01T10:00:00Z'
        '400':
          description: Email already registered
          content:
            application/json:
              example:
                detail: Email already registered
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
  /api/auth/me:
    get:
      tags:
      - authentication
      summary: Get Current User
      description: Get information about the currently authenticated user
      operationId: get_current_user_info_api_auth_me_get
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                email: john.doe@example.com
                name: John Doe
                role: sender
                phone: '+1234567890'
                created_at: '2024-01-01T10:00:00Z'
                updated_at: '2024-01-01T10:00:00Z'
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
      security:
      - HTTPBearer: []
  /api/auth/refresh:
    post:
      tags:
      - authentication
      summary: Refresh Access Token
      description: Get a new access token using refresh token
      operationId: refresh_token_api_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/TokenRefresh'
              title: Token Data
              description: Refresh token data
            example:
              refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        required: true
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 1800
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
  /api/auth/logout:
    post:
      tags:
      - authentication
      summary: User Logout
      description: Logout user (client should discard tokens)
      operationId: logout_api_auth_logout_post
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema: {}
              example:
                message: Logged out successfully
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
      security:
      - HTTPBearer: []
...
  /api/auth/change-password:
    post:
      tags:
      - authentication
      summary: Change Password
      description: Change password for the authenticated user
      operationId: change_password_api_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PasswordChange:
      properties:
        current_password:
          type: string
          minLength: 1
          title: Current Password
        new_password:
          type: string
          maxLength: 100
          minLength: 6
          title: New Password
      type: object
      required:
      - current_password
      - new_password
      title: PasswordChange
    RefreshTokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - expires_in
      title: RefreshTokenResponse
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/UserResponse'
      type: object
      required:
      - access_token
      - expires_in
      - user
      title: Token
...
    TokenRefresh:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: TokenRefresh
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone
        password:
          type: string
          maxLength: 100
          minLength: 6
          title: Password
      type: object
      required:
      - email
      - name
      - role
      - password
      title: UserCreate
    UserLogin:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - email
      - name
      - role
      - id
      - created_at
      - updated_at
      title: UserResponse
    UserRole:
      type: string
      enum:
      - admin
      - courier
      - sender
      - recipient
      title: UserRole
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint
security:
- BearerAuth: []
...